Made a fix to the code to ensure that only the owner of an image can edit or delete it.

+++ src/main/java/ImageHoster/controller/ImageController.java
@@ -46,10 +46,21 @@ public class ImageController {
     //Here a list of tags is added in the Model type object
     //this list is then sent to 'images/image.html' file and the tags are displayed
     @RequestMapping("/images/{imageId}/{title}")
-    public String showImage(@PathVariable("imageId") Integer imageId, @PathVariable("title") String title, Model model) {
-        Image image = imageService.getImageObject(imageId, title);
+    public String showImage(@PathVariable("imageId") Integer imageId, @PathVariable("title") String title, Model model) throws NullPointerException{
+        Image image = imageService.getImage(imageId);
         model.addAttribute("image", image);
-        model.addAttribute("tags", image.getTags());
+
+        try {
+            List<Tag> tags = image.getTags();
+            if (tags.isEmpty()) {
+                tags.add (new Tag());
+            }
+            model.addAttribute("tags", tags);
+        }
+        catch (NullPointerException e) {
+            e.printStackTrace();
+            model.addAttribute("image", "");
+        }
         return "images/image";
     }
 
@@ -92,13 +103,27 @@ public class ImageController {
     //The method first needs to convert the list of all the tags to a string containing all the tags separated by a comma and then add this string in a Model type object
     //This string is then displayed by 'edit.html' file as previous tags of an image
     @RequestMapping(value = "/editImage")
-    public String editImage(@RequestParam("imageId") Integer imageId, Model model) {
+    //Editing this method to ensure that the user session is managed
+    public String editImage(@RequestParam("imageId") Integer imageId, Model model, HttpSession session) {
         Image image = imageService.getImage(imageId);
+        User currentUser = (User) session.getAttribute("loggeduser");
+        User imageOwner = image.getUser();
 
         String tags = convertTagsToString(image.getTags());
-        model.addAttribute("image", image);
-        model.addAttribute("tags", tags);
-        return "images/edit";
+
+        if (currentUser.getId() == imageOwner.getId()) {
+            model.addAttribute("image", image);
+            model.addAttribute("tags", tags);
+            return "images/edit";
+        }
+        else {
+            image.setaccessDeniedError("Only the owner of the image can edit the image");
+            model.addAttribute("image", image);
+            model.addAttribute("tags", tags);
+            return "images/image";
+        }
+
+
     }
 
     //This controller method is called when the request pattern is of type 'images/edit' and also the incoming request is of PUT type
@@ -141,9 +166,22 @@ public class ImageController {
     //The method calls the deleteImage() method in the business logic passing the id of the image to be deleted
     //Looks for a controller method with request mapping of type '/images'
     @RequestMapping(value = "/deleteImage", method = RequestMethod.DELETE)
-    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId) {
-        imageService.deleteImage(imageId);
-        return "redirect:/images";
+    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId, Model model, HttpSession session) {
+        User currentUser = (User) session.getAttribute("loggeduser");
+        Image currentImage = imageService.getImage(imageId);
+        User imageOwner = currentImage.getUser();
+
+        if (currentUser.getId() == imageOwner.getId()) {
+            imageService.deleteImage(imageId);
+            return "redirect:/images";
+        }
+        else {
+            String tags = convertTagsToString(currentImage.getTags());
+            model.addAttribute("tags", currentImage.getTags());
+            currentImage.setaccessDeniedError("Only the owner of the image can delete the image");
+            model.addAttribute("image", currentImage);
+            return "images/image";
+        }
     }
